+++ care/src/main/java/com/animal/api/auth/controller/AuthController.java
@@ -33,6 +33,10 @@ public class AuthController {
 
 	/**
 	 * 일반 , 보호시설 사용자의 통합 로그인 메서드
+	 * 
+	 * @param dto 로그인 폼
+	 * @session 로그인 정보 저장
+	 * @return 로그인 한 사용자의 정보
 	 */
 	@PostMapping("/login")
 	public ResponseEntity<?> login(@RequestBody LoginRequestDTO dto, HttpSession session) {
