+++ care/src/main/java/com/animal/api/auth/mapper/AuthMapper.java
@@ -7,18 +7,13 @@ import com.animal.api.auth.model.vo.UserVO;
 
 @Mapper
 public interface AuthMapper {
-    /**
-     * 아이디로 사용자 조회 (유저, 보호소 공통)
-     */
-    UserVO findUserById(String id);
+	
+	// 아이디로 사용자 조회 (유저, 보호소 공통)
+	UserVO findUserById(String id);
 
-    /**
-     * 보호소 상세 정보 조회 (보호소 전용)
-     */
-    ShelterVO findShelterByUserIdx(int userIdx);
+    // 보호소 상세 정보 조회 (보호소 전용)
+	ShelterVO findShelterByUserIdx(int userIdx);
 
-    /**
-     * 로그인 성공 시 마지막 로그인 시간 갱신
-     */
-    void updateLastLoginAt(int userIdx);
+	// 로그인 성공 시 마지막 로그인 시간 갱신
+	void updateLastLoginAt(int userIdx);
 }
