+++ care/src/main/java/com/animal/api/auth/model/request/LoginRequestDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class LoginRequestDTO {
-	
+
 	private String Id;
 	private String password;
 }
